dangnhap

+++ web_timviec/routes/api.php
@@ -20,15 +20,17 @@ use App\Http\Controllers\WorkplaceController;
 |
 */
 
-Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
-    return $request->user();
-});
+// Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
+//     return $request->user();
+// });
 //api reset password
 // Route::post('/reset-password', 'AuthController@sendMailCandidate');
 // Route::put('/reset-password/{token}', 'AuthController@resetCandidate');
 // Route::post('/reset-password', 'AuthController@sendMailEmployer');
 // Route::put('/reset-password/{token}', 'AuthController@resetEmployer');
 
+//api logout
+Route::middleware('auth:sanctum')->post('/logout', [AuthController::class, 'logout']);
 //api auth
 Route::post('/loginCandidate', [AuthController::class,'loginCandidate']);
 Route::post('/registerCandidate', [AuthController::class,'registerCandidate']);
@@ -38,10 +40,36 @@ Route::post('/loginAdmin', [AuthController::class,'loginAdmin']);
 Route::post('/registerAdmin', [AuthController::class,'registerAdmin']);
 Route::post('/logout',[AuthController::class,'logout']);
 
-//api lĩnh vực
-Route::get('/getIndustry',[IndustryController::class,'getAllIndustry']);
-Route::post('/addIndustry',[IndustryController::class,'addIndustry']);
-Route::put('/updateIndustry/{id}',[IndustryController::class,'updateIndustry']);
-Route::delete('/deleteIndustry/{id}',[IndustryController::class,'deleteIndustry']);
-Route::post('industry/search',[IndustryController::class,'searchIndustry']);
+
+
+//api nơi làm việc
+
+Route::middleware('role:admin')->group(function () {
+   //api lĩnh vực
+    Route::get('/admin/getIndustry',[IndustryController::class,'getAllIndustry']);
+    Route::post('/admin/addIndustry',[IndustryController::class,'addIndustry']);
+    Route::put('/admin/updateIndustry/{id}',[IndustryController::class,'updateIndustry']);
+    Route::delete('/admin/deleteIndustry/{id}',[IndustryController::class,'deleteIndustry']);
+    Route::post('/admin/searchIndustry',[IndustryController::class,'searchIndustry']);
+
+    //api nơi làm việc
+    Route::get('/admin/getWorkplace',[IndustryController::class,'getAllWorkplace']);
+    Route::post('/admin/addWorkplace',[IndustryController::class,'addWorkplace']);
+    Route::put('/admin/updateWorkplace/{id}',[IndustryController::class,'updateWorkplace']);
+    Route::delete('/admin/deleteWorkplace/{id}',[IndustryController::class,'deleteWorkplace']);
+    Route::post('/admin/searchWorkplace',[IndustryController::class,'searchWorkplace']);
+});
+
+Route::middleware('role:employer')->group(function () {
+    
+    
+    Route::get('/employer/dashboard', [EmployerController::class, 'dashboard']);
+        // Các route khác cho employer
+});
+
+Route::middleware('role:api')->group(function () {
+    Route::get('/candidate/dashboard', [CandidateController::class, 'dashboard']);
+    
+   
+});
